MGM: Use uint32_t for job ids and number of copy jobs
This is required to adapt to the equivalent change in XRootD, and
will require building EOS against XRootD 5.7 or later.

See also https://github.com/xrootd/xrootd/issues/2235

+++ mgm/convert/ConversionJob.hh
@@ -63,7 +63,7 @@ public:
   //! @param source         the source url of the current job
   //! @param destination    the destination url of the current job
   //----------------------------------------------------------------------------
-  void BeginJob(uint16_t jobNum, uint16_t jobTotal,
+  void BeginJob(uint32_t jobNum, uint32_t jobTotal,
                 const XrdCl::URL* source,
                 const XrdCl::URL* destination) override
   {
@@ -78,7 +78,7 @@ public:
   //! @param bytesProcessed bytes processed by the current job
   //! @param bytesTotal     total number of bytes to be processed
   //----------------------------------------------------------------------------
-  void JobProgress(uint16_t jobNum, uint64_t bytesProcessed,
+  void JobProgress(uint32_t jobNum, uint64_t bytesProcessed,
                    uint64_t bytesTotal) override
   {
     mBytesProcessed = bytesProcessed;
@@ -88,7 +88,7 @@ public:
   //----------------------------------------------------------------------------
   //! Determine whether the job should be cancelled
   //----------------------------------------------------------------------------
-  bool ShouldCancel(uint16_t jobNum) override
+  bool ShouldCancel(uint32_t jobNum) override
   {
     return mCancel;
   }
