+++ mgm/proc/user/File.cc
@@ -966,7 +966,7 @@ ProcCommand::File()
                 url_trg.SetPath(target);
                 properties.Set("source", url_src);
                 properties.Set("target", url_trg);
-                properties.Set("sourceLimit", (uint16_t) 1);
+                properties.Set("sourceLimit", (uint32_t) 1);
                 properties.Set("chunkSize", (uint32_t)(4 * 1024 * 1024));
                 properties.Set("parallelChunks", (uint8_t) 1);
                 XrdCl::CopyProcess lCopyProcess;
